[stylistic change]issues fixed refered by Faizan
